wave2fbo: create just one vbo and reuse it

+++ lib/signal/signal/buffer.cpp
@@ -71,7 +71,7 @@ void MonoBuffer::
 }
 
 
-float MonoBuffer::
+double MonoBuffer::
         start() const
 {
     return (sample_offset_/sample_rate_).asFloat();
